lib: restructure `classification`
- Moved all resumption related things to `classification` proper.
