+++ crates/rsonpath-lib/src/engine/main.rs
@@ -11,14 +11,17 @@
 #[cfg(feature = "head-skip")]
 use super::head_skipping::{CanHeadSkip, HeadSkip};
 use super::Compiler;
+#[cfg(feature = "head-skip")]
+use crate::classification::ResumeClassifierState;
 use crate::classification::{
     quotes::{classify_quoted_sequences, QuoteClassifiedIterator},
     structural::{classify_structural_characters, Structural, StructuralIterator},
-    ClassifierWithSkipping,
 };
 use crate::debug;
 use crate::engine::depth::Depth;
 use crate::engine::error::EngineError;
+#[cfg(feature = "tail-skip")]
+use crate::engine::tail_skipping::TailSkip;
 use crate::engine::{Engine, Input};
 use crate::query::automaton::{Automaton, State};
 use crate::query::error::CompilerError;
@@ -74,6 +77,19 @@ fn empty_query<R: QueryResult>(bytes: &AlignedBytes<alignment::Page>) -> R {
     result
 }
 
+#[cfg(feature = "tail-skip")]
+macro_rules! Classifier {
+    () => {
+        TailSkip<'b, Q, I>
+    };
+}
+#[cfg(not(feature = "tail-skip"))]
+macro_rules! Classifier {
+    () => {
+        I
+    };
+}
+
 struct Executor<'q, 'b> {
     depth: Depth,
     state: State,
@@ -118,10 +134,12 @@ impl<'q, 'b> Executor<'q, 'b> {
     fn run_and_exit<R: QueryResult>(mut self, result: &mut R) -> Result<(), EngineError> {
         let quote_classifier = classify_quoted_sequences(self.bytes.relax_alignment());
         let structural_classifier = classify_structural_characters(quote_classifier);
-        self.run_on_subtree(
-            &mut ClassifierWithSkipping::new(structural_classifier),
-            result,
-        )?;
+        #[cfg(feature = "tail-skip")]
+        let mut classifier = TailSkip::new(structural_classifier);
+        #[cfg(not(feature = "tail-skip"))]
+        let mut classifier = structural_classifier;
+
+        self.run_on_subtree(&mut classifier, result)?;
 
         self.verify_subtree_closed()
     }
@@ -132,7 +150,7 @@ impl<'q, 'b> Executor<'q, 'b> {
         R: QueryResult,
     >(
         &mut self,
-        classifier: &mut ClassifierWithSkipping<'b, Q, I>,
+        classifier: &mut Classifier!(),
         result: &mut R,
     ) -> Result<(), EngineError> {
         while let Some(event) = self.next_event.or_else(|| classifier.next()) {
@@ -163,7 +181,7 @@ impl<'q, 'b> Executor<'q, 'b> {
 
     fn handle_colon<Q, I, R>(
         &mut self,
-        classifier: &mut ClassifierWithSkipping<'b, Q, I>,
+        classifier: &mut Classifier!(),
         idx: usize,
         result: &mut R,
     ) -> Result<(), EngineError>
@@ -212,7 +230,7 @@ impl<'q, 'b> Executor<'q, 'b> {
 
     fn handle_comma<Q, I, R>(
         &mut self,
-        classifier: &mut ClassifierWithSkipping<'b, Q, I>,
+        classifier: &mut Classifier!(),
         idx: usize,
         result: &mut R,
     ) -> Result<(), EngineError>
@@ -236,7 +254,7 @@ impl<'q, 'b> Executor<'q, 'b> {
 
     fn handle_opening<Q, I, R>(
         &mut self,
-        classifier: &mut ClassifierWithSkipping<'b, Q, I>,
+        classifier: &mut Classifier!(),
         idx: usize,
         result: &mut R,
     ) -> Result<(), EngineError>
@@ -328,7 +346,7 @@ impl<'q, 'b> Executor<'q, 'b> {
 
     fn handle_closing<Q, I>(
         &mut self,
-        classifier: &mut ClassifierWithSkipping<'b, Q, I>,
+        classifier: &mut Classifier!(),
         idx: usize,
     ) -> Result<(), EngineError>
     where
@@ -489,20 +507,25 @@ impl<'q, 'b> CanHeadSkip<'b> for Executor<'q, 'b> {
         &mut self,
         next_event: Structural,
         state: State,
-        classifier: &mut ClassifierWithSkipping<'b, Q, I>,
+        structural_classifier: I,
         result: &'r mut R,
-    ) -> Result<(), EngineError>
+    ) -> Result<ResumeClassifierState<'b, Q>, EngineError>
     where
         Q: QuoteClassifiedIterator<'b>,
         R: QueryResult,
         I: StructuralIterator<'b, Q>,
     {
+        #[cfg(feature = "tail-skip")]
+        let mut classifier = TailSkip::new(structural_classifier);
+        #[cfg(not(feature = "tail-skip"))]
+        let mut classifier = structural_classifier;
+
         self.state = state;
         self.next_event = Some(next_event);
 
-        self.run_on_subtree(classifier, result)?;
+        self.run_on_subtree(&mut classifier, result)?;
         self.verify_subtree_closed()?;
 
-        Ok(())
+        Ok(classifier.stop())
     }
 }
